move tests related to aggregation checks to own test script
